tests/wpt/tests/wai-aria/role/list-roles.html

<!doctype html>
<html>
<head>
  <title>List-related Role Verification Tests</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/resources/testdriver.js"></script>
  <script src="/resources/testdriver-vendor.js"></script>
  <script src="/resources/testdriver-actions.js"></script>
  <script src="/wai-aria/scripts/aria-utils.js"></script>
</head>
<body>

<p>Tests <a href="https://w3c.github.io/aria/#list">list</a> and related roles.</p>

<div role="list" data-testname="first simple list" data-expectedrole="list" class="ex">
  <div role="listitem" data-testname="first simple listitem" data-expectedrole="listitem" class="ex">x</div>
  <div role="listitem" data-testname="last simple listitem" data-expectedrole="listitem" class="ex">x</div>
</div>

<!--
  CORE-AAM requires that, for elements with roles not contained in the
  required context, user agents must ignore the role token and return the
  computed role as if the ignored role token had not been included.
  See https://w3c.github.io/core-aam/#roleMappingComputedRole
-->
<div role="listitem" data-testname="orphan div with listitem role" class="ex-generic">x</div>
<p role="listitem" data-testname="orphan p with listitem role" data-expectedrole="paragraph" class="ex">x</p>

<script>
AriaUtils.verifyRolesBySelector(".ex");
AriaUtils.verifyGenericRolesBySelector(".ex-generic");
</script>

</body>
</html>
