LayoutTests/imported/w3c/web-platform-tests/wai-aria/role/list-roles-expected.txt

Tests list and related roles.

x
x
x
x


PASS first simple list
PASS first simple listitem
PASS last simple listitem
FAIL orphan p with listitem role assert_equals: <p role="listitem" data-testname="orphan p with listitem role" data-expectedrole="paragraph" class="ex">x</p> expected "paragraph" but got "listitem"
